Struct iup::callback::button::KeyStates

pub struct KeyStates(_);

The state of mouse buttons and some keyboard buttons.

Methods

impl KeyStates

fn is_shift(&self) -> bool

Whether this state have a SHIFT key pressed.

fn is_control(&self) -> bool

Whether this state have a CONTROL key pressed.

fn is_alt(&self) -> bool

Whether this state have a ALT key pressed.

fn is_sys(&self) -> bool

Whether this state have the system key pressed.

The system key in Windows is the Windows key and in Mac is the Apple key.

fn is_double(&self) -> bool

Whether this state have the specified button in the callback doubly pressed.

fn is_button1(&self) -> bool

Whether this state have the left mouse button pressed.

fn is_button2(&self) -> bool

Whether this state have the middle mouse button pressed.

fn is_button3(&self) -> bool

Whether this state have the right mouse button pressed.

fn is_button4(&self) -> bool

Whether this state have the additional mouse button 1 pressed.

fn is_button5(&self) -> bool

Whether this state have the additional mouse button 2 pressed.