Trait iup::callback::button::MotionCb

pub trait MotionCb where Self: Element + 'static {
    fn set_motion_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self, i32, i32, KeyStates)> { ... }
    fn remove_motion_cb(&mut self) -> Option<Box<Callback<(Self, i32, i32, KeyStates)>>> { ... }
}

Action generated when the mouse moves.

The i32 parameters are the x,y position in the canvas where the event has occurred, in pixels.

The KeyStates parameter is the state of the mouse buttons and some keyboard keys at

Provided Methods

fn set_motion_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self, i32, i32, KeyStates)>

fn remove_motion_cb(&mut self) -> Option<Box<Callback<(Self, i32, i32, KeyStates)>>>