Module iup::callback::callbacks

Traits

Action
Action generated when the element is activated. Affects each element differently.

ActionCb
Action generated when the element is activated. Affects each element differently.

CaretCb
Action generated when the caret/cursor position is changed.

CloseCb
Called just before a dialog is closed when the user clicks the close button of the title bar or an equivalent action.

DestroyCb
Called right before an element is destroyed.

DropFilesCb
Action called when a file is dropped into the control.

EnterWindowCb
Action generated when the mouse enters the native element.

GetFocusCb
Action generated when an element is given keyboard focus.

HelpCb
Action generated when the user press F1 at a control.

KillFocusCb
Action generated when an element loses keyboard focus.

LeaveWindowCb
Action generated when the mouse leaves the native element.

MapCb
Called right after an element is mapped and its attributes updated in Widget::map.

MoveCb
Called after the widget is moved, see it's documentation for more details.

ResizeCb
Action generated when the canvas or dialog size is changed.

SpinCb
Action generated when a spin button is pressed.

UnmapCb
Called right before an element is unmapped.

ValueChangedCb
Usually called after the value of a control changed.

Functions

remove_close_cb
Removes a previosly set up close callback.

remove_idle
Removes a previosly set up idle_action callback.

set_close_cb
Action generated when IUP closes (i.e. at the end of with_iup).

set_idle
Action generated when there are no events or messages to be processed.