Trait iup::callback::callbacks::GetFocusCb

pub trait GetFocusCb where Self: Element + 'static {
    fn set_getfocus_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self,)> { ... }
    fn remove_getfocus_cb(&mut self) -> Option<Box<Callback<(Self,)>>> { ... }
}

Action generated when an element is given keyboard focus.

This callback is called after the KillFocusCb of the element that loosed the focus. The IupGetFocus (TODO) function during the callback returns the element that loosed the focus.

Provided Methods

fn set_getfocus_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self,)>

fn remove_getfocus_cb(&mut self) -> Option<Box<Callback<(Self,)>>>