Struct iup::clipboard::Clipboard

pub struct Clipboard(_);

An element that allows access to the clipboard.

You can use only one for the entire application because it does not store any data inside. Or you can simply create and drop every time you need to copy or paste.

See the clipboard module documentation for more details on how the system clipboard works.

Other platform-dependent attributes can be found on the IUP Clipboard Documentation.

Ownership

The clipboard must be manually destroyed, thus for the user safety it returns a guarded object on the new constructor.

Please refer to the crate level documentation of IUP-Rust (the main doc page) for details on ownership of elements.

Methods

impl Clipboard

fn new() -> Guard<Clipboard>

Creates a new clipboard operarator.

fn clear(&mut self) -> Self

Clears any data on the clipboard.

fn add_format<S: Into<String>>(&mut self, format: S) -> Self

Register a custom format for clipboard data given its name.

fn has_text(&mut self) -> bool

Informs if there is a text available at the clipboard.

fn set_text<S: Into<String>>(&mut self, text: S) -> Self

Copy text into the clipboard.

fn text(&mut self) -> Option<String>

Paste text off the clipboard.

fn has_image(&mut self) -> bool

Informs if there is a image available at the clipboard.

fn set_image<I: ImageElement>(&mut self, image: &I) -> Self

Copy text into the clipboard.

fn has_data<S: Into<String>>(&mut self, format: S) -> bool

Informs if there is data of the specified format available at the clipboard.

fn set_data<S, D>(&mut self, format: S, data: D) -> Self where S: Into<String>, D: AsRef<[u8]>

Copy data from the specified format into the clipboard.

fn data<S: Into<String>>(&mut self, format: S) -> Option<Vec<u8>>

Paste data from the specified format off the clipboard.