Trait iup::control::list::ListAction

pub trait ListAction where Self: Element {
    fn set_action<F>(&mut self, cb: F) -> Self where F: Callback<(Self, String, u32, ListItemState)> { ... }
    fn remove_action(&mut self) -> Option<Box<Callback<(Self, String, u32, ListItemState)>>> { ... }
}

Action generated when the state of an item in the list is changed. Also provides information on the changed item.

The String parameter is the text of the changed item. The u32 parameter is the number of the changed item starting at 1. The ListItemState parameter is whether the item was selected or deselected.

The ListItemState::Deselected is simulated internally by IUP in all systems. If you add or remove items to/from the list and you count on the ListItemState::Deselected value, then after adding/removing items set the VALUE attribute to ensure proper ListItemState::Deslected value.

Provided Methods

fn set_action<F>(&mut self, cb: F) -> Self where F: Callback<(Self, String, u32, ListItemState)>

fn remove_action(&mut self) -> Option<Box<Callback<(Self, String, u32, ListItemState)>>>