Trait iup::dialog::dialog::TrayClickCb

pub trait TrayClickCb where Self: Element + 'static {
    fn set_trayclick_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self, MouseButton, MouseButtonState, bool)> { ... }
    fn remove_trayclick_cb(&mut self) -> Option<Box<Callback<(Self, MouseButton, MouseButtonState, bool)>>> { ... }
}

[Windows and GTK Only]: Called right after the mouse button is pressed or released over the tray icon.

The fourth callback parameter is a bool indicating whether the button was double pressed.

CallbackReturn::Close will be processed.

Provided Methods

fn set_trayclick_cb<F>(&mut self, cb: F) -> Self where F: Callback<(Self, MouseButton, MouseButtonState, bool)>

fn remove_trayclick_cb(&mut self) -> Option<Box<Callback<(Self, MouseButton, MouseButtonState, bool)>>>