Function iup::element::global

pub fn global<S: Into<String>>(name: S) -> Option<String>

Returns an attribute value from the global environment.

The value can be returned from the driver or from the internal storage.

This function’s return value is not necessarily the same one used by the application to set the attribute’s value.