Struct iup::element::Handle

pub struct Handle(_);

An object that can wrap any IUP element.

The handle also provides implementation for traits that mayn't be implemented for the contained element, so be careful when using it.

Methods

impl Handle

fn from_named<S: Into<String>>(name: S) -> Option<Handle>

Constructs from a name associated with a element handle (with Element::add_handle_name or LED).

fn try_downcast<E: Element>(self) -> Result<E, Handle>

Converts this handle object into a element object if they are compatible.