Module iup::led

LED (Dialog-specification language) functionalities.

LED is a dialog-specification language whose purpose is not to be a complete programming language, but rather to make dialog specification simpler than in C. Additionally it allows users to easily edit your application layout from external files without touching any source code.

In LED, attributes and expressions follow this form:

elem = element[attribute1=value1,attribute2=value2,...](...expression...)

The names of the elements must not contain the “iup” prefix. Attribute values are always interpreted as strings, but they need to be in quotes (“…”) only when they include spaces. The “IUP_” prefix must not be added to the names of the attributes and predefined values. Expressions contain parameters for creating the element.

In LED there is no distinction between upper and lower case, except for attribute names.

Also there is no optional parameters, in arrays at least one parameter must exist.

To simply view a LED file objects use the LED Viewer application called IupView, in the applications included in the distribution. Pre-compiled binaries are available at the Download.

You need to check out the IUP documentation for each control to see their respective function signatures in LED.

Note: Using LED may allow you to create controls not yet implemented in iup-rust and that's fine. Use a Handle to have access to controls created from LED.

Functions

load
Compiles a LED specification from a file.

load_buffer
Compiles a LED specification from a string.