Struct iup::timer::Timer

pub struct Timer(_);

A timer which periodically invokes a callback when the time is up.

Ownership

The timer must be manually destroyed, thus for the user safety it returns a guarded object on the new constructor.

Please refer to the crate level documentation of IUP-Rust (the main doc page) for details on ownership of elements.

Methods

impl Timer

fn new() -> Guard<Timer>

Constructs a timer.

fn time(&self) -> Option<u32>

Gets the set time interval in milliseconds or None if not set.

fn set_time(&mut self, time: u32) -> Self

Sets the time interval in milliseconds.

In Windows the minimum value is 10ms.

fn run(&mut self) -> Self

Starts the timer.

Does nothing if the TIME attribute is not set i.e. set_time.

If you have multiple threads start the timer in the main thread.

fn stop(&mut self) -> Self

Stops the timer.

fn is_running(&self) -> bool

Returns the current timer state.