Trait iup::element::ConvertXYToPos

pub trait ConvertXYToPos: Element {
    fn convert_xy_to_pos(&self, x: i32, y: i32) -> Option<i32> { ... }
}

Provided Methods

fn convert_xy_to_pos(&self, x: i32, y: i32) -> Option<i32>

Converts a x,y coordinate in an item position in the container.

The x,y coordinates are relative to the left corner and top corner of the element.

This have a different effect for each control it is applied to, check their documentation.